+++ greencoffee/src/main/java/com/mauriciotogneri/greencoffee/ScreenCapture.java
@@ -3,9 +3,9 @@ package com.mauriciotogneri.greencoffee;
 import android.app.Activity;
 import android.graphics.Bitmap;
 import android.graphics.Bitmap.CompressFormat;
-import android.support.test.InstrumentationRegistry;
-import android.support.test.runner.lifecycle.ActivityLifecycleMonitorRegistry;
-import android.support.test.runner.lifecycle.Stage;
+import androidx.test.InstrumentationRegistry;
+import androidx.test.runner.lifecycle.ActivityLifecycleMonitorRegistry;
+import androidx.test.runner.lifecycle.Stage;
 import android.view.View;
 
 import java.io.File;