+++ greencoffee/src/main/java/com/mauriciotogneri/greencoffee/ScreenshotProvider.java
@@ -6,7 +6,7 @@ import java.text.SimpleDateFormat;
 import java.util.Date;
 import java.util.Locale;
 
-import static android.support.test.InstrumentationRegistry.getTargetContext;
+import static androidx.test.InstrumentationRegistry.getTargetContext;
 
 public interface ScreenshotProvider
 {