+++ greencoffee/src/main/java/com/mauriciotogneri/greencoffee/interactions/ActionableData.java
@@ -1,15 +1,15 @@
 package com.mauriciotogneri.greencoffee.interactions;
 
-import android.support.test.espresso.DataInteraction;
-import android.support.test.espresso.UiController;
-import android.support.test.espresso.ViewAction;
-import android.support.test.espresso.ViewAssertion;
+import androidx.test.espresso.DataInteraction;
+import androidx.test.espresso.UiController;
+import androidx.test.espresso.ViewAction;
+import androidx.test.espresso.ViewAssertion;
 import android.view.View;
 import android.widget.TextView;
 
 import org.hamcrest.Matcher;
 
-import static android.support.test.espresso.matcher.ViewMatchers.isAssignableFrom;
+import static androidx.test.espresso.matcher.ViewMatchers.isAssignableFrom;
 
 public class ActionableData extends ActionableObject
 {